fix(config): split disk snapshot methods to separate file (lint: file size)
