+++ internal/api/core/runtime_manager.go
@@ -651,35 +651,6 @@ func (r *APIRuntime) GetPosterManager() poster.PosterManagerInterface {
 
 // Server lifecycle methods (ServerCtx, Shutdown) are defined in server_lifecycle.go.
 
-// SetInitialConfigs publishes the runtime config and disk-origin snapshot as a
-// single atomic unit under reloadMu. The runtime config is the env-overridden
-// cfg handlers serve; the disk snapshot is the pre-env cfg persisted to disk,
-// used by the config-update service for redacted-secret preservation.
-func (r *APIRuntime) SetInitialConfigs(runtimeCfg, diskCfg *config.Config) {
-	r.reloadMu.Lock()
-	defer r.reloadMu.Unlock()
-	if runtimeCfg != nil && r.deps != nil && r.deps.CoreDeps != nil {
-		r.deps.CoreDeps.SetConfig(runtimeCfg)
-	}
-	if diskCfg != nil {
-		r.diskConfigSnapshot = diskCfg.Clone()
-	} else {
-		r.diskConfigSnapshot = nil
-	}
-}
-
-// DiskConfigSnapshot returns the cached pre-env disk config, or nil if no
-// snapshot has been published yet (callers should fall back to config.Load).
-func (r *APIRuntime) DiskConfigSnapshot() *config.Config {
-	r.reloadMu.RLock()
-	cfg := r.diskConfigSnapshot
-	r.reloadMu.RUnlock()
-	if cfg == nil {
-		return nil
-	}
-	return cfg.Clone()
-}
-
 // SetConfig sets the full application config and rebuilds the APIConfig snapshot.
 // This is a convenience method for test setup. Production code should use
 // APIRuntime.ReplaceReloadable() instead.
